Refactor Flatten/Dedupe metadatas into utils, fix merge issues
